Add helping classes for commands and property changes
__Notes on `NotifyPropertyChangedImpl`:__

The class `NotifyPropertyChangedImpl` implements the interface `INotifyPropertyChanged` which is necessary for the View to notice changes made in the ViewModel.

With this base class it is easier for the ViewModel classes to raise such PropertyChanged events without repeating the necessary boilerplate code each time.

__Notes on `DelegateCommand`:__

The class `DelegateCommand` implements the interface `ICommand` which is necessary for the View to execute commands when e.g. clicking on a GUI element.

With this class, the ViewModel has only to provide the (parameterless) function to execute when calling the command. Optionally, you can also provide a function for checking if the command can be executed.

Using the `DelegateCommand` class, you do not need to implement the full `ICommand` interface for each commands but can focus on the actual execute methods.