Move NotifyPropertyChangedImpl to separate project Common
__Note:__

The project Common can be used for classes that are used e.g. both by the model and the view model. An example is the moved class NotifyPropertyChangedImpl.

By moving NotifyPropertyChangedImpl it is possible by the model to directly use a implementation of INotifyPropertyChanged which is used for the view to get updates of the model's values (which is accessible via the view model). It would also be possible to build an intermediate step inside the view model to forward the model's update, however this would add redundant code only for updating e.g. the currently played song.