Refactor the timer variables for the current song
* Remove nullable and initialize with 0 for clarity
* Remove null checks that are not necessary anymore

__Note:__

By removing the nullability, the initial values are always 0 and thus the CurrentSongView always displays 0:00 for the current and maximum time when no song is loaded yet. This also removes a GUI glitch where the CurrentSongView grows in height after a song (and thus the timer labels) have been set.