Rename SimplePlaylist to Playlist
__Note:__

The initial thought was to create a class MultiPlaylist for playing multiple playlists in a row. However, it seems to be easier to use another structure for this future feature.

When using MultiPlaylist, you have to keep track of the current index of both the sub playlist (SimplePlaylist) and the current song within this sub playlist. When playing the next song, as an example, this makes the logic difficult to understand. Thus, only use simple playlists with straight-forward logic.