Navigate to different views when selecting another button on the navigation sidebar
* Add dummy views to be shown when selecting the different buttons
* Add logic for changing the view via MainWindowViewModel

__Note:__

The implementation method for enabling only one view at each time was made with boolean variables for each view that can be shown. Thus, only one boolean variable must be set to `true` while the others have to be set to `false.`

It would also be possible to have a ContentControl element in MainWindow and only change the current view element. However, this would require the ViewModel to have access to the View project to store and switch to the different View instances. To avoid this coupling, the approach with the boolean variables has been followed.