Navigate through playlist directories and show their contents
* Add and remove playlist directories
* Show sub directories of a playlist directories
* Show playlist files inside a playlist directories
* Move base paths of the app from AppConfigurator to AppConfig and make them `public static readonly`
* Add dialog for giving the playlist directory to add a name

__Notes to the DialogService:__

The DialogService resides in an own project Dialog. It is meant to be called when a dialog should be opened. This call needs access on both a ViewModel and a View object and thus breaks a strict MVVM implementation. To keep this necessary coupling at a minimum, the separate project was created to only have the connection of ViewModel and View in there.

Note that the ViewModel needs access to the Dialog project but the Dialog project does not refer to ViewModel (which would not be possible, anyway, because of circular dependencies). The ViewModel instance which is given to the show dialog method is expected to be an object and not a concrete ViewModel type.