Never hide the 'Back' button on the playlist GUI elements
__Note:__

Previously the 'Back' button was collapsed when there was no current directory path set in the PlaylistsView. This behavior is correctly when navigating through sub directories and going back to the root directory for playlists where the current directory path is indeed set to null.

However, when showing the current selected playlist, there always has to be a 'Back' button, even if this playlist is placed in the root directory (which causes the current directory path to be null). So, the Visibility binding is removed now. Changes in the view model are not necessary because this case is already taken into account.