Set 'UseLayoutRounding' to 'True' for sharp graphics
Before, the images or the separators sometimes had blurred edges because the edge fell in the middle of a device pixel ('anti-aliasing'). Now, the edges will be placed exactly on a device pixel for showing sharp edges.

The layout rounding is activated in the main window because it is inherited by children then.

For more information see:

https://docs.microsoft.com/en-us/dotnet/api/system.windows.frameworkelement.uselayoutrounding?view=windowsdesktop-6.0

https://stackoverflow.com/questions/45506447/wpf-listbox-separator-is-shown-with-different-thickness