Sanitize the song title when exporting the song as file
If not sanitizing the song title, the application crashes when characters that are invalid for file names are included within the title of a song that should be exported to a directory. As an example, a song including the character '?' in its title could not be exported since the question mark is an invalid character for Windows files.